Rework overridden import sub
We want two things to happen on import
  * Log::Report needs to me imported into the appropriate caller, and
  * Dancer2::Plugin (via Exporter::Tiny) needs to DWIM to initialize the
    plugin.

The target for Log::Report import differs depending on the version of
Dancer2::Plugin. The plugin rewrite for in Dancer2 > 0.166001 has one less
caller in the stack. All args from 'use' are passed to Log::Reports
import code.

Then the overridden import method is called, with the appropriate "into"
for the caller.